// src/lib/orval/generated/model/projectsControllerFindByOrganization200.ts
/**
 * Generated by orval v7.13.1 🍺
 * Do not edit manually.
 * Gradegy API
 * Interactive documentation for the Gradegy API
 * OpenAPI spec version: 1.0
 */
import type { Project } from './project';
import type { ProjectsControllerFindByOrganization200Meta } from './projectsControllerFindByOrganization200Meta';

export type ProjectsControllerFindByOrganization200 = {
  data?: Project[];
  meta?: ProjectsControllerFindByOrganization200Meta;
};