// src/lib/orval/generated/model/categoriesControllerFindByProject200.ts
/**
 * Generated by orval v7.13.1 🍺
 * Do not edit manually.
 * Gradegy API
 * Interactive documentation for the Gradegy API
 * OpenAPI spec version: 1.0
 */
import type { Category } from './category';
import type { CategoriesControllerFindByProject200Meta } from './categoriesControllerFindByProject200Meta';

export type CategoriesControllerFindByProject200 = {
  data?: Category[];
  meta?: CategoriesControllerFindByProject200Meta;
};