// src/lib/orval/generated/model/subcategoriesControllerFindAllByCategory200.ts
/**
 * Generated by orval v7.13.1 🍺
 * Do not edit manually.
 * Gradegy API
 * Interactive documentation for the Gradegy API
 * OpenAPI spec version: 1.0
 */
import type { Subcategory } from './subcategory';
import type { SubcategoriesControllerFindAllByCategory200Meta } from './subcategoriesControllerFindAllByCategory200Meta';

export type SubcategoriesControllerFindAllByCategory200 = {
  data?: Subcategory[];
  meta?: SubcategoriesControllerFindAllByCategory200Meta;
};